Fix segmenter import for CLI
